fix: hack to make metadata not required
The metadata.{name,namespace} keys are populated at build time by the location of the resource in the configuration path, so we don't need users to explicitly specify it. However, this is a hack! It also makes nested metadata keys optional, when they might be required. I should fix this.

Actually, most resource schemas already don't mark metadata as required (including the built-in Kubernetes ones). But Cilium's resource schemas do, which makes this annoying.